+++ src/main/java/io/ejangs/docsa/global/mongo/deletion/util/MongoIdsCollector.java
@@ -1,10 +1,10 @@
-package io.ejangs.docsa.global.mongoDeleteSystem.util;
+package io.ejangs.docsa.global.mongo.deletion.util;
 
 import io.ejangs.docsa.domain.branch.entity.Branch;
 import io.ejangs.docsa.domain.commit.dao.mongodb.CommitBlockSequenceRepository;
 import io.ejangs.docsa.domain.commit.entity.Commit;
 import io.ejangs.docsa.domain.save.entity.Save;
-import io.ejangs.docsa.global.mongoDeleteSystem.dto.MongoIdsDto;
+import io.ejangs.docsa.global.mongo.deletion.dto.MongoIdsDto;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
