+++ src/test/java/io/ejangs/docsa/mongoDeleteSystem/MongoDeleteRetryServiceTest.java
@@ -16,9 +16,9 @@ import io.ejangs.docsa.domain.doc.util.DocTestUtils;
 import io.ejangs.docsa.domain.save.dao.mongodb.SaveContentRepository;
 import io.ejangs.docsa.domain.user.dao.mysql.UserRepository;
 import io.ejangs.docsa.domain.user.entity.User;
-import io.ejangs.docsa.global.mongoDeleteSystem.app.MongoDeleteRetryService;
-import io.ejangs.docsa.global.mongoDeleteSystem.dao.mysql.MongoDeleteFailureRepository;
-import io.ejangs.docsa.global.mongoDeleteSystem.entity.MongoDeleteFailure;
+import io.ejangs.docsa.global.mongo.deletion.app.MongoDeleteRetryService;
+import io.ejangs.docsa.global.mongo.deletion.dao.mysql.MongoDeleteFailureRepository;
+import io.ejangs.docsa.global.mongo.deletion.entity.MongoDeleteFailure;
 import java.time.Duration;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
